Refactor regex for default Workflow URLs (redux)
THIS is the regex pattern intended for GH-277.

Previous commit message:

Switch from static base pattern of `logic.azure.com` to a regex
OR pattern to permit either of `logic.azure.com` or
`*.azure-api.net` as has been observed in the wild.

This also has the side effect of resolving potential
CodeQL alerts raised in PR GH-275.

refs GH-262