Expose `TeamsMessage` interface to support mocking
Export the `TeamsMessage` interface type to better support
abstracting/mocking the functionality of this library by
dependent projects.

The fields of the interface are intentionally not exported so
as to help prevent future changes/improvements from breaking
client code.

This change is considered a non-breaking change as the support
for creating message types compliant with this interface is
unchanged.

refs GH-285